+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/OrderRowBuilder.cs
@@ -147,7 +147,7 @@
             return this;
         }
 
-        public int GetDiscountPercent()
+        public decimal GetDiscountPercent()
         {
             return _discountPercent;
         }
