+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/ShippingFeeBuilder.cs
@@ -125,7 +125,7 @@
             return this;
         }
 
-        public int GetDiscountPercent()
+        public decimal GetDiscountPercent()
         {
             return _discountPercent;
         }
